Refactor to multiple files
