+++ .changeset/slow-spiders-fly.md
@@ -1,0 +1,8 @@
+---
+'astro': patch
+'@astrojs/prism': patch
+'@astrojs/mdx': patch
+'@astrojs/markdown-remark': patch
+---
+
+Refactor `@astrojs/mdx` and `@astrojs/markdown-remark` to use `@astrojs/prism` instead of duplicating the code
