+++ packages/astro-prism/Prism.astro
@@ -1,5 +1,5 @@
 ---
-import { runHighlighterWithAstro } from './dist/internal';
+import { runHighlighterWithAstro } from './dist/index';
 
 export interface Props {
 	class?: string;
