+++ packages/astro-prism/src/index.ts
@@ -1,2 +1,3 @@
 // @ts-expect-error
 export { default as Prism } from '../Prism.astro';
+export { runHighlighterWithAstro } from './highlighter';
