+++ packages/astro-prism/src/plugin.ts
@@ -1,7 +1,4 @@
-import Prism from 'prismjs';
-import loadLanguages from 'prismjs/components/index.js';
-
-function addAstro() {
+export function addAstro(Prism: typeof import('prismjs')) {
 	if (Prism.languages.astro) {
 		return;
 	}
@@ -187,43 +184,4 @@ function addAstro() {
 		}
 		walkTokens(env.tokens);
 	});
-}
-
-const languageMap = new Map([['ts', 'typescript']]);
-
-export function runHighlighterWithAstro(lang: string | undefined, code: string) {
-	let classLanguage = `language-${lang}`;
-
-	if (!lang) {
-		lang = 'plaintext';
-	}
-
-	const ensureLoaded = (language: string) => {
-		if (language && !Prism.languages[language]) {
-			loadLanguages([language]);
-		}
-	};
-
-	if (languageMap.has(lang)) {
-		ensureLoaded(languageMap.get(lang)!);
-	} else if (lang === 'astro') {
-		ensureLoaded('typescript');
-		addAstro();
-	} else {
-		ensureLoaded('markup-templating'); // Prism expects this to exist for a number of other langs
-		ensureLoaded(lang);
-	}
-
-	if (lang && !Prism.languages[lang]) {
-		// eslint-disable-next-line no-console
-		console.warn(`Unable to load the language: ${lang}`);
-	}
-
-	const grammar = Prism.languages[lang];
-	let html = code;
-	if (grammar) {
-		html = Prism.highlight(code, grammar, lang);
-	}
-
-	return { classLanguage, html };
 }
